bot-xp/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
contourpy==1.2.1
cycler==0.12.1
discord==2.3.2
discord.py==2.4.0
fonttools==4.53.1
frozenlist==1.4.1
idna==3.8
kiwisolver==1.4.5
matplotlib==3.9.2
multidict==6.0.5
numpy==2.1.0
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.16.0
yarl==1.9.4
